% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amphora-package.R
\docType{package}
\name{amphora-package}
\alias{amphora}
\alias{amphora-package}
\title{amphora: Manipulate Clinical Ontologies}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Amphora was inititally created to serve as an
    extension of the chariot R package to create interactive
    Reingold-Tilford tree diagrams in D3.js. It has now been extended to
    include the reading, parsing, and authoring of OWL/RDF files for
    knowledge representation in the clinical medicine domain. Wrappers to
    install [ROBOT](http://robot.obolibrary.org/) and use it for format
    conversions are also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meerapatelmd/amphora}
  \item Report bugs at \url{https://github.com/meerapatelmd/amphora/issues}
}

}
\author{
\strong{Maintainer}: Meera Patel, MD \email{patelmeeray@gmail.com} (\href{https://orcid.org/0000-0002-3401-3948}{ORCID})

}
\keyword{internal}
