% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{create_classification}
\alias{create_classification}
\title{FUNCTION_TITLE}
\usage{
create_classification(
  class_hierarchy = list(`CANCER ANATOMY` = c("NAACCR_Laterality"), `CANCER CLINICAL
    PERFORMANCE` = c("ICDO3_Behavior"), `CANCER GRADES` = c("CNS_Tumors"), `CANCER
    STAGES` = c("AnnArbor_HodgkinLymphoma", "Lugano_NonHodgkinLymphoma",
    "NAACCR_Clinical_Stage", "NAACCR_Path_Stage"), `CANCER TYPE` = c("ICD10",
    "ICDO3_Histology", "ICDO3_Top_Topography", "ICDO3_Topography"), DRUGS =
    c("BY RXNORM INGREDIENT/GLUCOCORTICOID",
    "BY RXNORM INGREDIENT/IMMUNOMODULATOR_IMIDE",
    "BY RXNORM INGREDIENT/IMMUNOMODULATOR_TNFa_INHIBITOR",     
    "BY RXNORM INGREDIENT/IMMUNOMODULATOR", "BY RXNORM INGREDIENT/NSAID",
    "BY RXNORM INGREDIENT/STATIN", "BY RXNORM INGREDIENT/VASOPRESSOR"), LABS =
    c("BLEEDING_TIME", "BLOOD_GASES", "BLOOD_TYPE", "CBC", "CREATINE_KINASE",
    "DIFFERENTIAL", "IMMUNOGLOBULIN_PANEL", "INFECTIOUS_DISEASE_DNA",
    "INFECTIOUS_DISEASE_RNA", "LACTATE_DEHYDROGENASE", "LIPID_PANEL", "METABOLIC_PANEL",
    "SINGLE_ORDER_LABS", "SPECIMEN_TYPE")),
  conn,
  standard_library_schema,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{class_hierarchy}{PARAM_DESCRIPTION, Default: list(\verb{CANCER ANATOMY} = c("NAACCR_Laterality"), \verb{CANCER CLINICAL PERFORMANCE} = c("ICDO3_Behavior"),
\verb{CANCER GRADES} = c("CNS_Tumors"), \verb{CANCER STAGES} = c("AnnArbor_HodgkinLymphoma",
"Lugano_NonHodgkinLymphoma", "NAACCR_Clinical_Stage",
"NAACCR_Path_Stage"), \verb{CANCER TYPE} = c("ICD10", "ICDO3_Histology",
"ICDO3_Top_Topography", "ICDO3_Topography"), DRUGS = c("BY RXNORM INGREDIENT/GLUCOCORTICOID",
"BY RXNORM INGREDIENT/IMMUNOMODULATOR_IMIDE", "BY RXNORM INGREDIENT/IMMUNOMODULATOR_TNFa_INHIBITOR",
"BY RXNORM INGREDIENT/IMMUNOMODULATOR", "BY RXNORM INGREDIENT/NSAID",
"BY RXNORM INGREDIENT/STATIN", "BY RXNORM INGREDIENT/VASOPRESSOR"),
LABS = c("BLEEDING_TIME", "BLOOD_GASES", "BLOOD_TYPE", "CBC",
"CREATINE_KINASE", "DIFFERENTIAL", "IMMUNOGLOBULIN_PANEL",
"INFECTIOUS_DISEASE_DNA", "INFECTIOUS_DISEASE_RNA", "LACTATE_DEHYDROGENASE",
"LIPID_PANEL", "METABOLIC_PANEL", "SINGLE_ORDER_LABS",
"SPECIMEN_TYPE"))}

\item{conn}{PARAM_DESCRIPTION}

\item{standard_library_schema}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[uuid]{UUIDgenerate}}
\code{\link[pg13]{append_table}}
\code{\link[tibble]{tibble}}
}
