% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemonc_to_protege.R
\name{excel_hemonc_class_hierarchy}
\alias{excel_hemonc_class_hierarchy}
\title{Write Excel of HemOnc's Class Hierarchy for Protege}
\usage{
excel_hemonc_class_hierarchy(
  conn,
  conn_fun,
  file = "hemonc_class_hierarchy.xlsx"
)
}
\description{
To print the HemOnc Class Hierarchy to the console to copy-and-paste into Protege's "owl:Thing", see \code{\link{print_hemonc_class_hierarchy}}.
}
