% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemonc_to_protege.R
\name{hemonc_triples}
\alias{hemonc_triples}
\title{Convert OMOP HemOnc to Triplets}
\usage{
hemonc_triples(
  conn,
  conn_fun,
  write_schema = "hemonc",
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\description{
Convert OMOP HemOnc to Triplets
}
