% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{install_robot}
\alias{install_robot}
\title{Install Robot}
\usage{
install_robot(
  jar_url = "https://github.com/ontodev/robot/releases/download/v1.7.2/robot.jar",
  sh_url = "https://raw.githubusercontent.com/ontodev/robot/master/bin/robot",
  path = "/usr/local/bin"
)
}
\description{
Downloads a jar and shell script to $PATH (\code{path}). Once complete, cd into
the path and run \verb{sudo chmod u+x robot} in the terminal to make ROBOT
executable. Learn more at http://robot.obolibrary.org/.
}
