% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{instantiate_standard_library}
\alias{instantiate_standard_library}
\title{Creating the Target Standard Library from OMOP Concepts}
\usage{
instantiate_standard_library(
  conn,
  standard_library_schema,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{standard_library_schema}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Filter the current OMOP vocabulary for the elements most
relevant to the RWD domain.
}
\details{
\enumerate{
\item Identify a valid Concept in the vocabulary schema. A hash
will be created for concepts that do not presently exist.
\item Provide the class it belongs to. If the Class does not
currently exist, it will be made and added to the Concept table.
\item Copy the Concept to the Standard Library.
\item Add the "Has concept" relationship to the Class
\item Add that "Concept of" relationship to the Concept
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{schema_exists}},\code{\link[pg13]{create_schema}},\code{\link[pg13]{send}}
\code{\link[SqlRender]{render}}
\code{\link[secretary]{typewrite}}
}
