% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOMOP_TABLES}
\alias{MOMOP_TABLES}
\title{mOMOP Tables}
\format{
An object of class \code{list} of length 2.
}
\usage{
MOMOP_TABLES
}
\description{
mOMOP in the OMOP Vocabulary format to reflect the
contrived class hierarchy.
}
\keyword{datasets}
