% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omop_concept-class.R
\docType{class}
\name{omop_concept-class}
\alias{omop_concept-class}
\alias{omop_concept}
\title{OMOP Concept Class (S4)}
\description{
This class allows to know the full OMOP concept representation in R Studio.
}
