% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_domain.R
\name{plot_atc_1st_class}
\alias{plot_atc_1st_class}
\title{Plot ATC 1st}
\usage{
plot_atc_1st_class(conn, vocab_schema = "omop_vocabulary", atc_1st_concept)
}
\description{
Plot the hierarchy of an ATC 1st Class.
}
