% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel-drug_domain2.R
\name{preview_hemonc_classification}
\alias{preview_hemonc_classification}
\title{Preview a HemOnc Class's Descendants}
\usage{
preview_hemonc_classification(
  conn,
  concept_class_obj,
  vocab_schema = "omop_vocabulary",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)
}
\description{
Preview a HemOnc Class's Descendants
}
