% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemonc_to_protege.R
\name{print_hemonc_class_hierarchy}
\alias{print_hemonc_class_hierarchy}
\title{Print Class Hierarchy for Protege}
\usage{
print_hemonc_class_hierarchy(conn, conn_fun)
}
\description{
Print the HemOnc Class Hierarchy to the console to copy-and-paste into Protege's "owl:Thing". To write an Excel to import using Cellphie instead, see \code{\link{excel_hemonc_class_hierarchy}}.
}
