% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot.R
\name{robot_convert}
\alias{robot_convert}
\title{Convert an Ontology Format}
\usage{
robot_convert(input, output)
}
\description{
Supported extensions include:
\itemize{
\item json
\item obo
\item ofn
\item omn
\item owl
\item owx
\item ttl
}

See \url{http://robot.obolibrary.org/convert} for more details.
}
