% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_add.R
\name{sl_add_classification}
\alias{sl_add_classification}
\title{FUNCTION_TITLE}
\usage{
sl_add_classification(
  class_hierarchy,
  conn,
  standard_library_schema,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{class_hierarchy}{PARAM_DESCRIPTION}

\item{conn}{PARAM_DESCRIPTION}

\item{standard_library_schema}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{c("query", "query")}},\code{\link[pg13]{append_table}}
\code{\link[SqlRender]{render}}
\code{\link[uuid]{UUIDgenerate}}
\code{\link[tibble]{tibble}}
}
