% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_add.R
\name{source_add_concept}
\alias{source_add_concept}
\title{FUNCTION_TITLE}
\usage{
source_add_concept(
  conn,
  source_concept_name,
  vocabulary_id = "Unspecified",
  concept_class_id = "Unspecified",
  standard_library_schema,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{source_concept_name}{PARAM_DESCRIPTION}

\item{vocabulary_id}{PARAM_DESCRIPTION, Default: 'Unspecified'}

\item{concept_class_id}{PARAM_DESCRIPTION, Default: 'Unspecified'}

\item{standard_library_schema}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{c("query", "query")}},\code{\link[pg13]{append_table}}
\code{\link[SqlRender]{render}}
\code{\link[uuid]{UUIDgenerate}}
\code{\link[tibble]{tibble}}
}
