% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_add.R
\name{source_add_map_to_target}
\alias{source_add_map_to_target}
\title{FUNCTION_TITLE}
\usage{
source_add_map_to_target(
  conn,
  standard_library_schema,
  source_concept_obj,
  target_concept_obj,
  verbose = TRUE,
  render_sql = TRUE,
  render_only = FALSE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{standard_library_schema}{PARAM_DESCRIPTION}

\item{source_concept_obj}{PARAM_DESCRIPTION}

\item{target_concept_obj}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{append_table}}
\code{\link[tibble]{tibble}}
}
