% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_library_concept-class.R
\docType{class}
\name{standard_library_concept-class}
\alias{standard_library_concept-class}
\alias{standard_library_concept}
\title{Standard Library Concept Class (S4)}
\description{
This class allows to know the full Standard Library concept representation in R Studio.
}
