% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UPTODATE_TABLES}
\alias{UPTODATE_TABLES}
\title{UpToDate Drug Tables}
\format{
An object of class \code{list} of length 2.
}
\usage{
UPTODATE_TABLES
}
\description{
Drugs in UpToDate in the OMOP Vocabulary format, classified based on
their assigned Pharmacologic Category in Lexicomp.
}
\keyword{datasets}
