% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_atc_classification.R
\name{preview_atc_classification}
\alias{preview_atc_classification}
\title{Preview an ATC Class's Descendants}
\usage{
preview_atc_classification(
  conn,
  concept_class_obj,
  vocab_schema = "omop_vocabulary",
  verbose = TRUE,
  render_sql = TRUE,
  sleep_time = 1
)
}
\description{
The ATC Classification includes the interface between the ATC 5th Concept Class and the RxNorm Ingredient Concept Class.

Due to the high amount of records, used to determine the appropriate range based on row counts per level to supply the optional \code{range} argument for \code{\link{plot_atc_classification}}, where this function is called again and can be optionally filtered on a numeric range before plotting.
}
